puma/pyproject.toml
[project]
name = "xuma"
version = "0.0.2"
description = "Pure Python implementation of xDS Unified Matcher API"
requires-python = ">=3.12"
license = "MIT"
readme = "README.md"
dependencies = ["google-re2>=1.1"]

[project.urls]
Repository = "https://github.com/mox-nexus/x.uma"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-benchmark>=5.0",
    "pyyaml>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/xuma"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
